@@ -51,33 +59,44 @@
         "@angular/platform-browser-dynamic": "6.0.0",
         "@angular/platform-server": "6.0.0",
         "@angular/router": "6.0.0",
-        "@compodoc/compodoc": "1.0.0-beta.9",
+        "@compodoc/compodoc": "1.1.5",
         "@types/jasmine": "2.6.2",
-        "@types/node": "^8.0.47",
-        "chalk": "1.1.3",
-        "codelyzer": "3.0.1",
-        "core-js": "2.5.5",
+        "@types/node": "10.12.0",
+        "chalk": "2.4.1",
+        "codelyzer": "4.4.4",
+        "core-js": "2.5.7",
+        "istanbul-instrumenter-loader": "3.0.1",
         "jasmine-core": "2.6.2",
-        "karma": "1.7.0",
-        "karma-chrome-launcher": "2.1.1",
-        "karma-jasmine": "1.1.0",
-        "karma-jasmine-html-reporter": "^0.2.2",
+        "karma": "3.1.1",
+        "karma-chrome-launcher": "2.2.0",
+        "karma-coverage-istanbul-reporter": "2.0.4",
+        "karma-jasmine": "1.1.2",
         "karma-sourcemap-loader": "0.3.7",
-        "karma-spec-reporter": "0.0.31",
-        "karma-webpack": "2.0.5",
-        "reflect-metadata": "0.1.10",
-        "rollup": "0.41.6",
-        "rollup-plugin-commonjs": "8.0.2",
-        "rollup-plugin-node-resolve": "3.0.0",
-        "rxjs": "6.1.0",
-        "rxjs-compat": "6.1.0",
-        "shelljs": "0.7.7",
-        "sorcery": "0.10.0",
-        "ts-loader": "2.1.0",
-        "tslint": "5.8.0",
+        "karma-spec-reporter": "0.0.32",
+        "karma-webpack": "4.0.0-rc.2",
+        "node-sass": "4.9.4",
+        "reflect-metadata": "0.1.12",
+        "rollup": "0.66.6",
+        "rollup-plugin-license": "0.7.0",
+        "rollup-plugin-node-resolve": "3.4.0",
+        "rollup-plugin-sourcemaps": "0.4.2",
+        "rollup-plugin-commonjs": "9.2.0",
+        "rxjs": "6.3.3",
+        "shelljs": "0.8.2",
+        "source-map-loader": "0.2.4",
+        "ts-loader": "5.2.2",
+        "tslint": "5.11.0",
         "typescript": "2.7.2",
-        "webpack": "3.8.1",
-        "yargs": "8.0.1",
-        "zone.js": "0.8.26"
-    }
+        "uglify-js": "3.4.9",
+        "webpack": "4.21.0",
+        "zone.js": "0.8.26",
+        "angular2-template-loader": "0.6.2",
+        "raw-loader": "0.5.1",
+        "css-loader": "1.0.0",
+        "sass-loader": "7.1.0",
+        "to-string-loader": "1.1.5",
+        "tsickle": "0.33.0",
+        "webpack-filter-warnings-plugin": "1.2.0"
+    },
+    "sideEffects": false
 }
