@@ -1,70 +1,103 @@
 'use strict';
 
-require('shelljs/global');
+const shell = require('shelljs');
 const chalk = require('chalk');
 
 const PACKAGE = `angular-auth-oidc-client`;
 const NPM_DIR = `dist`;
-const MODULES_DIR = `${NPM_DIR}/modules`;
+const ESM2015_DIR = `${NPM_DIR}/esm2015`;
+const ESM5_DIR = `${NPM_DIR}/esm5`;
+const FESM2015_DIR = `${NPM_DIR}/fesm2015`;
+const FESM5_DIR = `${NPM_DIR}/fesm5`;
 const BUNDLES_DIR = `${NPM_DIR}/bundles`;
+const OUT_DIR = `${NPM_DIR}/package`;
+const OUT_DIR_ESM5 = `${NPM_DIR}/package/esm5`;
 
-echo(`Start building...`);
+shell.echo(`Start building...`);
 
-rm(`-Rf`, `${NPM_DIR}/*`);
-mkdir(`-p`, `./${MODULES_DIR}`);
-mkdir(`-p`, `./${BUNDLES_DIR}`);
+shell.rm(`-Rf`, `${NPM_DIR}/*`);
+shell.mkdir(`-p`, `./${ESM2015_DIR}`);
+shell.mkdir(`-p`, `./${ESM5_DIR}`);
+shell.mkdir(`-p`, `./${FESM2015_DIR}`);
+shell.mkdir(`-p`, `./${FESM5_DIR}`);
+shell.mkdir(`-p`, `./${BUNDLES_DIR}`);
+shell.mkdir(`-p`, `./${OUT_DIR}`);
 
 /* TSLint with Codelyzer */
 // https://github.com/palantir/tslint/blob/master/src/configs/recommended.ts
 // https://github.com/mgechev/codelyzer
-echo(`Start TSLint`);
-exec(`tslint --project ./tsconfig.json --type-check ./src/**/*.ts`);
-echo(chalk.green(`TSLint completed`));
+shell.echo(`Start TSLint`);
+shell.exec(`tslint -p tsconfig.json -t stylish src/**/*.ts`);
+shell.echo(chalk.green(`TSLint completed`));
 
-/* AoT compilation: ES2015 sources */
-echo(`Start AoT compilation`);
-if (exec(`ngc -p tsconfig-build.json`).code !== 0) {
-    echo(chalk.red(`Error: AoT compilation failed`));
-    exit(1);
+shell.cp(`-Rf`, [`src`, `*.ts`, `*.json`], `${OUT_DIR}`);
+
+/* Try to process scss files  */
+shell.echo(`Try to process scss files`);
+if (shell.exec(`node-sass -r ${OUT_DIR} -o ${OUT_DIR}`).code === 0) {
+    shell.rm(`-Rf`, `${OUT_DIR}/**/*.scss`);
+    shell.ls(`${OUT_DIR}/**/*.css`).forEach(function(file) {
+        shell.mv(file, file.replace('.css', '.scss'));
+    });
 }
-echo(chalk.green(`AoT compilation completed`));
 
-/* Creates bundles: ESM/ES5 and UMD bundles */
-echo(`Start bundling`);
-echo(`Rollup package`);
-exec(`rollup -i ${NPM_DIR}/index.js -o ${MODULES_DIR}/${PACKAGE}.js --sourcemap`, { silent: true });
-exec(`node scripts/map-sources -f ${MODULES_DIR}/${PACKAGE}.js`);
-
-echo(`Downleveling ES2015 to ESM/ES5`);
-cp(`${MODULES_DIR}/${PACKAGE}.js`, `${MODULES_DIR}/${PACKAGE}.es5.ts`);
-exec(`tsc ${MODULES_DIR}/${PACKAGE}.es5.ts --target es5 --module es2015 --noLib --sourceMap`, { silent: true });
-exec(`node scripts/map-sources -f ${MODULES_DIR}/${PACKAGE}.es5.js`);
-rm(`-f`, `${MODULES_DIR}/${PACKAGE}.es5.ts`);
-
-echo(`Run Rollup conversion on package`);
-if (exec(`rollup -c rollup.config.js --sourcemap`).code !== 0) {
-    echo(chalk.red(`Error: Rollup conversion failed`));
-    exit(1);
+/* AoT compilation */
+shell.echo(`Start AoT compilation`);
+if (shell.exec(`ngc -p ${OUT_DIR}/tsconfig-build.json`).code !== 0) {
+    shell.echo(chalk.red(`Error: AoT compilation failed`));
+    shell.exit(1);
 }
-exec(`node scripts/map-sources -f ${BUNDLES_DIR}/${PACKAGE}.umd.js`);
+shell.echo(chalk.green(`AoT compilation completed`));
 
-echo(`Minifying`);
-cd(`${BUNDLES_DIR}`);
-exec(
-    `uglifyjs -c --screw-ie8 --comments -o ${PACKAGE}.umd.min.js --source-map ${PACKAGE}.umd.min.js.map --source-map-include-sources ${PACKAGE}.umd.js`,
-    { silent: true }
-);
-exec(`node ../../scripts/map-sources -f ${PACKAGE}.umd.min.js`);
-cd(`..`);
-cd(`..`);
+shell.echo(`Copy ES2015 for package`);
+shell.cp(`-Rf`, [`${NPM_DIR}/src/`, `${NPM_DIR}/*.js`, `${NPM_DIR}/*.js.map`], `${ESM2015_DIR}`);
 
-echo(chalk.green(`Bundling completed`));
+/* BUNDLING PACKAGE */
+shell.echo(`Start bundling`);
+shell.echo(`Rollup package`);
+if (shell.exec(`rollup -c rollup.es.config.js -i ${NPM_DIR}/${PACKAGE}.js -o ${FESM2015_DIR}/${PACKAGE}.js`).code !== 0) {
+    shell.echo(chalk.red(`Error: Rollup package failed`));
+    shell.exit(1);
+}
 
-rm(`-Rf`, `${NPM_DIR}/*.js`);
-rm(`-Rf`, `${NPM_DIR}/*.js.map`);
-rm(`-Rf`, `${NPM_DIR}/src/**/*.js`);
-rm(`-Rf`, `${NPM_DIR}/src/**/*.js.map`);
+shell.echo(`Produce ESM5/FESM5 versions`);
+shell.exec(`ngc -p ${OUT_DIR}/tsconfig-build.json --target es5 -d false --outDir ${OUT_DIR_ESM5} --sourceMap`);
+shell.cp(`-Rf`, [`${OUT_DIR_ESM5}/src/`, `${OUT_DIR_ESM5}/*.js`, `${OUT_DIR_ESM5}/*.js.map`], `${ESM5_DIR}`);
+if (shell.exec(`rollup -c rollup.es.config.js -i ${OUT_DIR_ESM5}/${PACKAGE}.js -o ${FESM5_DIR}/${PACKAGE}.js`).code !== 0) {
+    shell.echo(chalk.red(`Error: FESM5 version failed`));
+    shell.exit(1);
+}
 
-cp(`-Rf`, [`package.json`, `LICENSE`, `README.md`], `${NPM_DIR}`);
+shell.echo(`Run Rollup conversion on package`);
+if (shell.exec(`rollup -c rollup.config.js -i ${FESM5_DIR}/${PACKAGE}.js -o ${BUNDLES_DIR}/${PACKAGE}.umd.js`).code !== 0) {
+    shell.echo(chalk.red(`Error: Rollup conversion failed`));
+    shell.exit(1);
+}
 
-echo(chalk.green(`End building`));
+shell.echo(`Minifying`);
+shell.cd(`${BUNDLES_DIR}`);
+if (
+    shell.exec(
+        `uglifyjs ${PACKAGE}.umd.js -c --comments -o ${PACKAGE}.umd.min.js --source-map "includeSources=true,content='${PACKAGE}.umd.js.map',filename='${PACKAGE}.umd.min.js.map'"`
+    ).code !== 0
+) {
+    shell.echo(chalk.red(`Error: Minifying failed`));
+    shell.exit(1);
+}
+shell.cd(`..`);
+shell.cd(`..`);
+
+shell.echo(chalk.green(`Bundling completed`));
+
+shell.rm(`-Rf`, `${NPM_DIR}/package`);
+shell.rm(`-Rf`, `${NPM_DIR}/*.js`);
+shell.rm(`-Rf`, `${NPM_DIR}/*.js.map`);
+shell.rm(`-Rf`, `${NPM_DIR}/src/**/*.js`);
+shell.rm(`-Rf`, `${NPM_DIR}/src/**/*.js.map`);
+shell.rm(`-Rf`, `${ESM2015_DIR}/src/**/*.d.ts`);
+
+shell.cp(`-Rf`, [`package.json`, `LICENSE`, `README.md`], `${NPM_DIR}`);
+
+shell.sed('-i', `"private": true,`, `"private": false,`, `./${NPM_DIR}/package.json`);
+
+shell.echo(chalk.green(`End building`));
