@@ -1,7 +1,10 @@
 // Karma configuration for Unit testing
 
+const path = require('path');
+const FilterWarningsPlugin = require('webpack-filter-warnings-plugin');
+
 module.exports = function(config) {
-    var configuration = {
+    const configuration = {
         // base path that will be used to resolve all patterns (eg. files, exclude)
         basePath: '',
 
@@ -15,6 +18,8 @@ module.exports = function(config) {
             require('karma-webpack'),
             require('karma-sourcemap-loader'),
             require('karma-spec-reporter'),
+            require('karma-coverage-istanbul-reporter'),
+            require('istanbul-instrumenter-loader'),
         ],
 
         // list of files / patterns to load in the browser
@@ -26,11 +31,12 @@ module.exports = function(config) {
         // preprocess matching files before serving them to the browser
         // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
         preprocessors: {
-            'spec.bundle.js': ['webpack'],
+            'spec.bundle.js': ['webpack', 'sourcemap'],
         },
 
         // webpack
         webpack: {
+            mode: 'development',
             resolve: {
                 extensions: ['.ts', '.js'],
             },
@@ -38,13 +44,43 @@ module.exports = function(config) {
                 rules: [
                     {
                         test: /\.ts/,
-                        loaders: ['ts-loader'],
+                        use: [{ loader: 'ts-loader' }, { loader: 'angular2-template-loader' }, { loader: 'source-map-loader' }],
                         exclude: /node_modules/,
+                    },
+                    {
+                        test: /\.html$/,
+                        use: 'raw-loader',
+                    },
+                    {
+                        test: /\.css$/,
+                        use: [{ loader: 'to-string-loader' }, { loader: 'css-loader' }],
+                    },
+                    {
+                        test: /\.scss$/,
+                        use: [{ loader: 'raw-loader' }, { loader: 'sass-loader' }],
+                    },
+                    {
+                        enforce: 'post',
+                        test: /\.ts/,
+                        use: [
+                            {
+                                loader: 'istanbul-instrumenter-loader',
+                                options: { esModules: true },
+                            },
+                        ],
+                        exclude: [/\.spec.ts/, /node_modules/],
                     },
                 ],
                 exprContextCritical: false,
             },
+            devtool: 'inline-source-map',
             performance: { hints: false },
+            /* Workaround for https://github.com/angular/angular/issues/21560 */
+            plugins: [
+                new FilterWarningsPlugin({
+                    exclude: /System.import/,
+                }),
+            ],
         },
 
         webpackServer: {
@@ -54,7 +90,13 @@ module.exports = function(config) {
         // test results reporter to use
         // possible values: 'dots', 'progress'
         // available reporters: https://npmjs.org/browse/keyword/karma-reporter
-        reporters: ['spec'],
+        reporters: ['spec', 'coverage-istanbul'],
+
+        coverageIstanbulReporter: {
+            reports: ['html', 'lcovonly'],
+            dir: path.join(__dirname, 'coverage'),
+            fixWebpackSourcePaths: true,
+        },
 
         // web server port
         port: 9876,
