@@ -1,38 +1,38 @@
 import commonjs from 'rollup-plugin-commonjs';
 import resolve from 'rollup-plugin-node-resolve';
+import sourcemaps from 'rollup-plugin-sourcemaps';
 
 const globals = {
     '@angular/core': 'ng.core',
     '@angular/common': 'ng.common',
     '@angular/router': 'ng.router',
     '@angular/common/http': 'ng.common.http',
-    'rxjs/Observable': 'Rx',
-    'rxjs/Observer': 'Rx',
-    'rxjs/BehaviorSubject': 'Rx',
-    'rxjs/operators': 'Rx.Observable.prototype',
-    'rxjs/observable/timer': 'Rx.Observable.prototype',
-
+    rxjs: 'rxjs',
+    'rxjs/operators': 'rxjs.operators',
     jsrsasign: 'jsrsasign',
     buffer: 'buffer',
 };
 
 export default {
-    entry: './dist/modules/angular-auth-oidc-client.es5.js',
-    dest: './dist/bundles/angular-auth-oidc-client.umd.js',
-    format: 'umd',
-    exports: 'named',
-    moduleName: 'ng.angularAuthOidcClient',
+    external: Object.keys(globals),
+    plugins: [resolve(), sourcemaps()],
+    onwarn: () => {
+        return;
+    },
+    output: {
+        format: 'umd',
+        name: 'ng.angularAuthOidcClient',
+        globals: globals,
+        sourcemap: true,
+        exports: 'named',
+        amd: { id: 'angular-library-starter' },
+    },
     plugins: [
         resolve(),
         commonjs({
             namedExports: {
-                'node_modules/jsrsasign/lib/jsrsasign.js': ['KJUR', 'KEYUTIL', 'hextob64u']
-            }
-        })
+                'node_modules/jsrsasign/lib/jsrsasign.js': ['KJUR', 'KEYUTIL', 'hextob64u'],
+            },
+        }),
     ],
-    external: Object.keys(globals),
-    globals: globals,
-    onwarn: () => {
-        return;
-    }
 };
